Refactor chat navigation and render hot paths
