+++ lib/ui/home/chat/files_preview.dart
@@ -1,5 +1,3 @@
-// ignore_for_file: parameter_assignments
-
 import 'dart:async';
 import 'dart:io';
 
@@ -49,6 +47,8 @@ import '../../provider/quote_message_provider.dart';
 import 'image_caption_input.dart';
 import 'image_editor.dart';
 
+part 'files_preview_model.dart';
+
 Future<void> showFilesPreviewDialog(
   BuildContext context,
   List<XFile> files,
@@ -58,129 +58,6 @@ Future<void> showFilesPreviewDialog(
     child: _FilesPreviewDialog(initialFiles: files.map(_File.auto).toList()),
   );
 }
-
-/// We need this view object to keep the value of file#length.
-sealed class _File {
-  const _File({required this.file});
-
-  factory _File.normal(String path) => _NormalFile._(file: File(path).xFile);
-
-  factory _File.image(File file, [ImageEditorSnapshot? snapshot]) =>
-      _ImageFile._(file: file.xFile, imageEditorSnapshot: snapshot);
-
-  factory _File.auto(XFile file) {
-    if (file.mimeType == null) {
-      e('mimeType is null');
-      file = file.withMineType();
-    }
-    if (file.isImage) {
-      return _ImageFile._(file: file);
-    } else if (kPlatformIsDarwin && file.isVideo) {
-      return _VideoFile._(file: file);
-    } else {
-      return _NormalFile._(file: file);
-    }
-  }
-
-  final XFile file;
-
-  bool get isImage => false;
-
-  bool get isVideo => false;
-
-  bool get isMedia => isImage || isVideo;
-
-  String get path => file.path;
-
-  String? get mimeType => file.mimeType;
-}
-
-class _NormalFile extends _File {
-  const _NormalFile._({required super.file});
-}
-
-class _ImageFile extends _File {
-  const _ImageFile._({required super.file, this.imageEditorSnapshot});
-
-  final ImageEditorSnapshot? imageEditorSnapshot;
-
-  @override
-  bool get isImage => true;
-}
-
-class _VideoMetadata {
-  _VideoMetadata({
-    required this.width,
-    required this.height,
-    required this.duration,
-  });
-
-  final int width;
-  final int height;
-  final int duration;
-}
-
-class _VideoFile extends _File {
-  _VideoFile._({required super.file}) {
-    _loadMetadata();
-  }
-
-  static Future<String?> _videoBlurHash(
-    (RootIsolateToken token, String videoPath) params,
-  ) async {
-    final (token, videoPath) = params;
-    BackgroundIsolateBinaryMessenger.ensureInitialized(token);
-    final bytes = await customVideoCompress.getByteThumbnail(videoPath);
-    final decodedImage = image.decodeImage(bytes!)!;
-    return BlurHash.encode(
-      image.copyResize(decodedImage, width: 100, maintainAspect: true),
-    ).hash;
-  }
-
-  final _metadataCompleter = Completer<_VideoMetadata?>();
-
-  final _blurHashCompleter = Completer<String?>();
-
-  Future<void> _loadMetadata() async {
-    try {
-      final mediaInfo = await customVideoCompress.getMediaInfo(file.path);
-      final duration = mediaInfo.duration ?? 0.0;
-      _metadataCompleter.complete(
-        _VideoMetadata(
-          width: mediaInfo.width ?? 0,
-          height: mediaInfo.height ?? 0,
-          duration: duration.floor(),
-        ),
-      );
-    } catch (error, stackTrace) {
-      e('failed to load video metadata', error, stackTrace);
-      _metadataCompleter.complete(null);
-    }
-
-    final stopwatch = Stopwatch()..start();
-    try {
-      final hash = await compute(_videoBlurHash, (
-        ServicesBinding.rootIsolateToken!,
-        file.path,
-      ));
-      _blurHashCompleter.complete(hash);
-    } catch (error, stackTrace) {
-      e('failed to load video thumbnail', error, stackTrace);
-      _blurHashCompleter.complete(null);
-    }
-
-    d('thumbnail cost: ${stopwatch.elapsedMilliseconds}ms');
-  }
-
-  @override
-  bool get isVideo => true;
-}
-
-typedef _ImageEditedCallback = void Function(_File, ImageEditorSnapshot);
-
-const _kDefaultArchiveName = 'Archive.zip';
-
-enum _TabType { image, files, zip }
 
 class _FilesPreviewDialog extends HookConsumerWidget {
   const _FilesPreviewDialog({required this.initialFiles});
